requirements.txt
python-docx
google-generativeai
python-dotenv